+++ RUFAS/routines/manure/manure_treatments/anaerobic_lagoon.py
@@ -42,14 +42,16 @@ class AnaerobicLagoon(BaseManureTreatment):
             (kg :math:`CH_4`/day).
 
         """
+        air_temperature = self._get_current_day_average_temperature_celsius()
+        storage_temperature = self._determine_outdoor_storage_temperature(air_temperature)
         # fmt: off
         methane_emission, methane_emission_from_degradable_volatile_solids = (
-            GasEmissionsCalculator.methane_emission_from_slurry_storage(
+            GasEmissionsCalculator.calculate_methane_emission_from_slurry_storage(
                 accumulated_liquid_manure_total_degradable_volatile_solids=(
                     accumulated_output.liquid_manure_total_degradable_volatile_solids),
                 accumulated_liquid_manure_total_non_degradable_volatile_solids=(
                     accumulated_output.liquid_manure_total_non_degradable_volatile_solids),
-                temp=self._get_current_day_average_temperature_celsius(),
+                storage_temperature=storage_temperature,
             )
         )
         # fmt: on
@@ -74,12 +76,14 @@ class AnaerobicLagoon(BaseManureTreatment):
             Update the `storage_ammonia` attribute of the accumulated output object.
 
         """
-        storage_ammonia_emission = GasEmissionsCalculator.storage_ammonia_emission(
+        air_temperature = self._get_current_day_average_temperature_celsius()
+        storage_temperature = self._determine_outdoor_storage_temperature(air_temperature)
+        storage_ammonia_emission = GasEmissionsCalculator.calculate_storage_ammonia_emission(
             num_animals=self._current_pen.num_animals,
             manure_total_ammoniacal_nitrogen=self._accumulated_output.liquid_manure_total_ammoniacal_nitrogen,
             manure_volume=self._accumulated_output.liquid_manure_daily_volume,
             manure_density=ManureConstants.LIQUID_MANURE_DENSITY,
-            temp=self._get_current_day_average_temperature_celsius(),
+            storage_temperature=storage_temperature,
         )
         daily_output.storage_ammonia = storage_ammonia_emission